lambda/docker/requirements.txt
boto3
botocore
python-dateutil
requests
